src/backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.6.2.post1
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==14.0
croniter==3.0.4
dagster==1.9.3
dagster-graphql==1.9.3
dagster-pipes==1.9.3
dagster-postgres==0.25.3
dagster-webserver==1.9.3
docstring_parser==0.16
fastapi==0.115.6
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gql==3.5.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-health-checking==1.62.3
h11==0.14.0
httptools==0.6.4
humanfriendly==10.0
idna==3.10
Jinja2==3.1.4
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
propcache==0.2.0
protobuf==4.25.5
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
structlog==24.4.0
tabulate==0.9.0
tomli==2.2.1
toposort==1.10
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
universal_pathlib==0.2.5
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchdog==5.0.3
watchfiles==1.0.0
websockets==14.1
xmltodict==0.14.2
yarl==1.18.0
